WIP: Separate Helsinki styles from the Kerrokantasi base styles
Kerrokantasi is planned to be a white label, standalone version
that could be used by any service provider. Therefore City of
Helsinki styles needs to be separated from the base styles of
Kerrokantasi. This commit separates the styles in the codebase
but does not affect the end build result of the styles.

The idea is to write default styles to new components and views
to the Kerrokantasi styles base. If there are some styles that
need any branding changes, these can be added to the `hel` folder.
It should be kept in mind when doing new changes that the app
should work without City of Helsinki customizations.

Here is a short list of what's changed:
* Move the base styles of the views and components from `hel` folder under `kerrokantasi` folder.
* Add temporary Sass variable values for kerrokantasi as a base.
* Override the default Kerrokantasi variables with Helsinki custom ones.
* Override some of the Kerrokantasi default styles with Helsinki custom ones.
* Add possibility to `src/index.js` to switch between white label styles and helsinki styles. Helsinki styles are now applied as default styles.
* Move alertify styles import from index.js to kerrokantasi.scss
* Write instructions to README how to apply custom theming